+++ src/schemas/trackerData.schema.ts
@@ -16,19 +16,6 @@ export const TrackerDataSchema = z.object({
   /** Base folder used to collect entries (if applicable). */
   basePath: z.string().optional(),
 
-  /**
-   * @deprecated The default intensity value for an entry.
-   */
-  defaultEntryIntensity: z.number(),
-  /**
-   * @deprecated The starting value for the intensity scale.
-   */
-  intensityScaleStart: z.number().optional(),
-  /**
-   * @deprecated The ending value for the intensity scale.
-   */
-  intensityScaleEnd: z.number().optional(),
-
   intensityConfig: IntensityConfigSchema,
   separateMonths: z.boolean().optional(),
   heatmapTitle: z.string().or(z.number()).optional(),